Refactored test code.
I moved all the test code to a single monolithic file.
Not having to deal with (possible) concurrency headaches is nice.
This simplifies things at the cost of performance.
I was getting random fails before and this approach seems more stable.
The performance loss is negligible for now but will obviously grow as the tests grow in number.
I am still not completely satisfied with this, but oh well.