Implemented logging with Winston everywhere.
Feels nice not to fly blind anymore!
Now all major CRUD operations are logged.
The logging, for the time being, is just outputted to
two log files (which I added to the .gitignore), 'errors.log', which contains only errors,
and 'general.log', which contains all messages.
Also made small but significant changes to many Promise-based functions
which were coded to reject on normal conditions instead of on an actual error.