Major refactoring of routes and front-end error/success handling.
- Modularized routes. No more 350-lines long unreadable spaghetti!
- Implemented semantic HTTP verbs. Now, instead of making a
POST to makePost/editPost/deletePost, we just make
a POST/PUT/DELETE to post, for example.
- Improved routes URL strings handling. Now we store all routes URLs
in a global variable, and consistently use it across the
entire application.
- Improved redirect URL strings handling. See above.
- Front-end no longer relies on HTML forms behavior, because you
can't PUT/DELETE from a form, for whatever reason. Now we consistently
use AJAX everywhere.
- We no longer use pug templates to convey error/success of an
operation. Instead, we consistently employ toasts everywhere. This not
only plays nicer with AJAX, but also looks better.
- We no longer use HTTP redirects because they don't play nice with
AJAX. Now we just send JSON with a 'redirect' property and call it a day.
- Updated tests to reflect changes.