Major REST design refactoring.
1) Made changes to enforce clean URLs: HTML files are no longer public.
I made this change because I realized that by setting the static folder
to 'public' with Express I was inadvertently letting unauthenticated
users get, for example, change_password.html by just doing
'http://localhost:3000/change_password.html', which is a) dangerous,
because I could mistakenly be serving sensitive html to the wrong people,
b) ugly, because the user shouldn't be aware of the underlying
implementation, whether it be an html file or what have you. Now, when
the user clicks on the 'Change password' button, he is redirected to
/password, which is MUCH nicer than /change_password.html.

2) Changed 'public' folder structure. Now only the subfolders /css, /js,
and /res are exposed to the public. Yes, I did keep /html inside /public
despite it not really being public anymore, I just find it comfortable
that way.

Also:
- Fixed broken css transitions.
- HTML files no longer hardcode the route to their css/js files.
They just use the filename, and let Express handle the routes for them.
This makes it much easier to move resources around without having to
update HTML files.
- Deleted pug files because they were phased out in favor of toasts + ajax.
- Updated eslint ignores to reflect file structure changes.
- Added a 404 route (how could I miss this for so long!?).