Major config management refactoring.
- Implemented so-called 'convention over configuration' by providing
sane defaults where it makes sense to do so. No more dumb .env
variables which don't ever change between deploys!

- Separated back-end config from front-end config, for security reasons.
Back-end config is stored in src/globals.js, and the config shared by
both the back-end and the front-end is in public/js/shared_globals.js.

- Made config checking more robust: When a critical setting is missing
from the back-end config, an error is logged and the process terminates
with error code 1. This is a nice way of telling the user "hey, your
.env file is missing critical variables."

Minor changes:
- 'error.pug' template is back because I found a legit use for it:
Showing an error page when the database is down or when we get a 404.
- Updated README to reflect changes.