Axed user avatar feature.
Decided not to implement the avatar feature after all for a couple of
reasons:

a) It's pretty hard to do well in mobile devices without turning the
UI into a mess (e.g. the post info, like user, post id, timestamp,
etc., turn into an eyesore when they flow around the avatar). This
would require a re-engineering of the post widget which is just not
worth it.

b) The project must be easily deployable to Heroku. This means that
avatar files can't be stored locally, because Heroku's filesystem
is ephemeral, which implies that we'd have to either 1) use the
database to store blob files (terrible idea), 2) use an external
service, like AWS S3. While 2) might be acceptable, I think it's not
worth the hassle just to store a couple of png files, considering you
have to enter credit card info and *could* be charged.